refactor: rename TourResponse - Tour refactor: move Tour from network
