+++ model/src/main/java/com/demal/model/data/entity/tours/Tour.kt
@@ -1,0 +1,8 @@
+package com.demal.model.data.entity.tours
+
+import com.demal.model.data.entity.tours.network.TourResponse
+
+data class Tour(
+    val tour: TourResponse,
+    val isFavorite: Boolean
+)
