+++ model/src/main/java/com/demal/model/data/entity/tours/Tours.kt
@@ -1,6 +1,5 @@
 package com.demal.model.data.entity.tours
 
 import com.demal.model.data.entity.AppStateEntity
-import com.demal.model.data.entity.tours.network.TourResponse
 
-data class Tours(val toursList : List<TourResponse>) : AppStateEntity
+data class Tours(val toursList : List<Tour>) : AppStateEntity
