+++ model/src/main/java/com/demal/model/data/entity/tours/network/TourResponse.kt
@@ -1,6 +1,18 @@
 package com.demal.model.data.entity.tours.network
 
 data class TourResponse(
-    val tour: Tour,
-    val isFavorite: Boolean
+    val id: Int,
+    val title: String?,
+    val subTitle: String?,
+    val place: String?,
+    val price: Int,
+    val country: String?,
+    val duration: Int,
+    val description: String?,
+    val travelersCount: Int,
+    val startDate: String?,
+    val categoryId: Int,
+    val createdAt: String?,
+    val category: CategoryResponse,
+    val images: List<ImageResponse>,
 )
