+++ repository/src/main/java/com/demal/repository/api/ApiService.kt
@@ -1,6 +1,6 @@
 package com.demal.repository.api
 
-import com.demal.model.data.entity.tours.network.Tour
+import com.demal.model.data.entity.tours.network.TourResponse
 import com.demal.model.data.entity.user.LoginRequest
 import com.demal.model.data.entity.user.LoginResponse
 import com.demal.model.data.entity.user.User
@@ -21,12 +21,12 @@ interface ApiService {
     fun getTours(
         @Query("sortBy") sortBy: SortBy,
         @Query("order") order: Order,
-    ): Deferred<List<Tour>>
+    ): Deferred<List<TourResponse>>
 
     @GET("api/$API_VERSION/tours/{id}")
     fun getTourById(
         @Path("id") id: Int,
-    ): Deferred<Tour>
+    ): Deferred<TourResponse>
 
     companion object {
         private const val API_VERSION = "v1"
