+++ repository/src/main/java/com/demal/repository/data_sources/RemoteDataSource.kt
@@ -1,6 +1,6 @@
 package com.demal.repository.data_sources
 
-import com.demal.model.data.entity.tours.network.Tour
+import com.demal.model.data.entity.tours.network.TourResponse
 import com.demal.model.data.entity.user.LoginRequest
 import com.demal.model.data.entity.user.LoginResponse
 import com.demal.model.data.entity.user.User
@@ -10,6 +10,6 @@ import com.demal.repository.types.SortBy
 interface RemoteDataSource {
     suspend fun login(auth: LoginRequest): LoginResponse
     suspend fun myUser(): User
-    suspend fun getTours(sortBy: SortBy, order: Order): List<Tour>
-    suspend fun getTourById(id: Int) : Tour
+    suspend fun getTours(sortBy: SortBy, order: Order): List<TourResponse>
+    suspend fun getTourById(id: Int) : TourResponse
 }
