+++ repository/src/main/java/com/demal/repository/repository/ToursRepository.kt
@@ -1,10 +1,10 @@
 package com.demal.repository.repository
 import com.demal.model.data.entity.tours.Tours
-import com.demal.model.data.entity.tours.network.TourResponse
+import com.demal.model.data.entity.tours.Tour
 import com.demal.repository.types.Order
 import com.demal.repository.types.SortBy
 
 interface ToursRepository {
     suspend fun getTours(sortBy: SortBy, order: Order) : Tours
-    suspend fun getTourById(id: Int) : TourResponse
+    suspend fun getTourById(id: Int) : Tour
 }
