+++ repository/src/main/java/com/demal/repository/repository/ToursRepositoryImpl.kt
@@ -1,8 +1,8 @@
 package com.demal.repository.repository
 
 import com.demal.model.data.entity.tours.Tours
-import com.demal.model.data.entity.tours.network.Tour
 import com.demal.model.data.entity.tours.network.TourResponse
+import com.demal.model.data.entity.tours.Tour
 import com.demal.repository.data_sources.RemoteDataSource
 import com.demal.repository.types.Order
 import com.demal.repository.types.SortBy
@@ -19,13 +19,13 @@ class ToursRepositoryImpl(
         mapTour(remoteDataSource.getTourById(id))
 
     // TODO get information about likes from Room to set isFavorite
-    private fun mapTours(toursList: List<Tour>): List<TourResponse> {
+    private fun mapTours(toursList: List<TourResponse>): List<Tour> {
         return toursList.map {
-            TourResponse(it, false)
+            Tour(it, false)
         }
     }
 
     // TODO get information about likes from Room to set isFavorite
-    private fun mapTour(tour: Tour) =
-        TourResponse(tour, false)
+    private fun mapTour(tour: TourResponse) =
+        Tour(tour, false)
 }
